requirements.txt
crewai>=0.28.0
langchain-groq>=0.1.3
python-dotenv>=1.0.0
streamlit>=1.35.0